+++ inst/doc/cast02-AOA-tutorial.R
@@ -3,10 +3,9 @@ knitr::opts_chunk$set(fig.width = 8.83)
 
 ## ---- message = FALSE, warning=FALSE------------------------------------------
 library(CAST)
-library(virtualspecies)
 library(caret)
-library(raster)
-library(sp)
+library(terra)
+#library(sp)
 library(sf)
 library(viridis)
 library(latticeExtra)
@@ -18,26 +17,62 @@ RMSE = function(a, b){
 }
 
 ## ---- message = FALSE, warning=FALSE------------------------------------------
-predictors <- stack(system.file("extdata","bioclim.grd",package="CAST"))
-spplot(stretch(predictors,0,1),col.regions=viridis(100))
+predictors <- rast(system.file("extdata","bioclim.grd",package="CAST"))
+plot(stretch(predictors,0,1),col=viridis(100))
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
-response <- generateSpFromPCA(predictors,
-                              means = c(3,1),sds = c(2,2), plot=F)$suitab.raster
+
+generate_random_response <- function(raster, predictornames =
+names(raster), seed = sample(seq(1000), 1)){
+  operands_1 = c("+", "-", "*", "/")
+  operands_2 = c("^1","^2")
+  
+  expression <- paste(as.character(predictornames, sep=""))
+  # assign random power to predictors
+  set.seed(seed)
+  expression <- paste(expression,
+                      sample(operands_2, length(predictornames),
+replace = TRUE),
+                      sep = "")
+  
+  # assign random math function between predictors (expect after the last one)
+  set.seed(seed)
+  expression[-length(expression)] <- paste(expression[-
+length(expression)],
+                                           sample(operands_1,
+length(predictornames)-1, replace = TRUE),
+                                           sep = " ")
+  print(paste0(expression, collapse = " "))
+  # collapse
+  e = paste0("raster$", expression, collapse = " ")
+  
+  response = eval(parse(text = e))
+  names(response) <- "response"
+  return(response)
+  
+}
+
+## ----message = FALSE, warning=FALSE-------------------------------------------
+
+response <- generate_random_response (predictors, seed = 10)
+plot(response,col=viridis(100),main="virtual response")
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 mask <- predictors[[1]]
 values(mask)[!is.na(values(mask))] <- 1
-mask <- rasterToPolygons(mask)
+mask <- st_as_sf(as.polygons(mask))
+
 set.seed(15)
-samplepoints <- spsample(mask,20,"random")
-spplot(response,col.regions=viridis(100),
-            sp.layout=list("sp.points", samplepoints, col = "red", first = FALSE, cex=2))
+samplepoints <- st_as_sf(st_sample(mask,20,"random"))
+
+plot(response,col=viridis(100))
+plot(samplepoints,col="red",add=T,pch=3)
+
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
-trainDat <- extract(predictors,samplepoints,df=TRUE)
-trainDat$response <- extract (response,samplepoints)
-trainDat <- trainDat[complete.cases(trainDat),]
+trainDat <- extract(predictors,samplepoints,na.rm=FALSE)
+trainDat$response <- extract(response,samplepoints,na.rm=FALSE, ID=FALSE)$response
+trainDat <- na.omit(trainDat)
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 set.seed(10)
@@ -53,9 +88,9 @@ print(model)
 plot(varImp(model,scale = F),col="black")
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
-prediction <- predict(predictors,model)
+prediction <- predict(predictors,model,na.rm=T)
 truediff <- abs(prediction-response)
-spplot(stack(prediction,response),main=c("prediction","reference"))
+plot(rast(list(prediction,response)),main=c("prediction","reference"))
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 AOA <- aoa(predictors, model)
@@ -66,43 +101,26 @@ print(AOA)
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 plot(AOA)
 
-## ----message = FALSE, warning=FALSE-------------------------------------------
-grid.arrange(
-  spplot(truediff,col.regions=viridis(100),main="true prediction error"),
-  spplot(AOA$DI,col.regions=viridis(100),main="DI"),
-  spplot(prediction, col.regions=viridis(100),main="prediction for AOA")+ spplot(AOA$AOA,col.regions=c("grey","transparent")), ncol=3)
+## ----message = FALSE, warning=FALSE,  fig.show="hold", out.width="30%"--------
+plot(truediff,col=viridis(100),main="true prediction error")
+plot(AOA$DI,col=viridis(100),main="DI")
+plot(prediction, col=viridis(100),main="prediction for AOA")
+plot(AOA$AOA,col=c("grey","transparent"),add=T)
 
-## ----clusteredpoints,message = FALSE, include=FALSE---------------------------
-#For a clustered sesign:
-csample <- function(x,n,nclusters,maxdist,seed){
-  set.seed(seed)
-  cpoints <- sp::spsample(x, n = nclusters, type="random")
-  result <- cpoints
-  result$clstrID <- 1:length(cpoints)
-  for (i in 1:length(cpoints)){
-    ext <- rgeos::gBuffer(cpoints[i,], width = maxdist)
-    newsamples <- sp::spsample(ext, n = (n-nclusters)/nclusters, 
-                               type="random")
-    newsamples$clstrID <- rep(i,length(newsamples))
-    result <- rbind(result,newsamples)
-    
-  }
-  result$ID <- 1:nrow(result)
-  return(result)
-}
+## ----message = FALSE, warning=FALSE-------------------------------------------
+set.seed(15)
+samplepoints <- clustered_sample(mask,75,15,radius=25000)
+
+plot(response,col=viridis(100))
+plot(samplepoints,col="red",add=T,pch=3)
+
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 
-samplepoints <- csample(mask,75,15,maxdist=0.20,seed=15)
-spplot(response,col.regions=viridis(100),
-            sp.layout=list("sp.points", samplepoints, col = "red", first = FALSE, cex=2))
-
-## ----message = FALSE, warning=FALSE-------------------------------------------
-
-trainDat <- extract(predictors,samplepoints,df=TRUE)
-trainDat$response <- extract (response,samplepoints)
-trainDat <- merge(trainDat,samplepoints,by.x="ID",by.y="ID")
-trainDat <- trainDat[complete.cases(trainDat),]
+trainDat <- extract(predictors,samplepoints,na.rm=FALSE)
+trainDat$response <- extract(response,samplepoints,na.rm=FALSE)$response
+trainDat <- data.frame(trainDat,samplepoints)
+trainDat <- na.omit(trainDat)
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 set.seed(10)
@@ -111,11 +129,11 @@ model_random <- train(trainDat[,names(predictors)],
                method="rf",
                importance=TRUE,
                trControl = trainControl(method="cv"))
-prediction_random <- predict(predictors,model_random)
+prediction_random <- predict(predictors,model_random,na.rm=TRUE)
 print(model_random)
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
-folds <- CreateSpacetimeFolds(trainDat, spacevar="clstrID",k=10)
+folds <- CreateSpacetimeFolds(trainDat, spacevar="parent",k=10)
 set.seed(15)
 model <- train(trainDat[,names(predictors)],
                  trainDat$response,
@@ -125,20 +143,19 @@ model <- train(trainDat[,names(predictors)],
                  trControl = trainControl(method="cv",index=folds$index))
   print(model)
   
-prediction <- predict(predictors,model)
+prediction <- predict(predictors,model,na.rm=TRUE)
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 AOA_spatial <- aoa(predictors, model)
 
 AOA_random <- aoa(predictors, model_random)
 
-## ----message = FALSE, warning=FALSE-------------------------------------------
-grid.arrange(spplot(AOA_spatial$DI,col.regions=viridis(100),main="DI"),
-  spplot(prediction, col.regions=viridis(100),main="prediction for AOA \n(spatial CV error applies)")+
-         spplot(AOA_spatial$AOA,col.regions=c("grey","transparent")),
-  spplot(prediction_random, col.regions=viridis(100),main="prediction for AOA \n(random CV error applies)")+
-         spplot(AOA_random$AOA,col.regions=c("grey","transparent")),
-ncol=3)
+## ----message = FALSE, warning=FALSE,  fig.show="hold", out.width="50%"--------
+plot(AOA_spatial$DI,col=viridis(100),main="DI")
+plot(prediction, col=viridis(100),main="prediction for AOA \n(spatial CV error applies)")
+plot(AOA_spatial$AOA,col=c("grey","transparent"),add=TRUE)
+plot(prediction_random, col=viridis(100),main="prediction for AOA \n(random CV error applies)")
+plot(AOA_random$AOA,col=c("grey","transparent"),add=TRUE)
 
 ## ---- message = FALSE, warning=FALSE------------------------------------------
 grid.arrange(plot(AOA_spatial) + ggplot2::ggtitle("Spatial CV"),
@@ -146,20 +163,24 @@ grid.arrange(plot(AOA_spatial) + ggplot2::ggtitle("Spatial CV"),
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 ###for the spatial CV:
-RMSE(values(prediction)[values(AOA_spatial$AOA)==1],values(response)[values(AOA_spatial$AOA)==1])
-RMSE(values(prediction)[values(AOA_spatial$AOA)==0],values(response)[values(AOA_spatial$AOA)==1])
+RMSE(values(prediction)[values(AOA_spatial$AOA)==1],
+     values(response)[values(AOA_spatial$AOA)==1])
+RMSE(values(prediction)[values(AOA_spatial$AOA)==0],
+     values(response)[values(AOA_spatial$AOA)==0])
 model$results
 
 ###and for the random CV:
-RMSE(values(prediction_random)[values(AOA_random$AOA)==1],values(response)[values(AOA_random$AOA)==1])
-RMSE(values(prediction_random)[values(AOA_random$AOA)==0],values(response)[values(AOA_random$AOA)==1])
+RMSE(values(prediction_random)[values(AOA_random$AOA)==1],
+     values(response)[values(AOA_random$AOA)==1])
+RMSE(values(prediction_random)[values(AOA_random$AOA)==0],
+     values(response)[values(AOA_random$AOA)==0])
 model_random$results
 
 ## ----message = FALSE, warning=FALSE-------------------------------------------
 AOA_calib <- calibrate_aoa(AOA_spatial,model,window.size = 5,length.out = 5, multiCV=TRUE,showPlot=FALSE)
 AOA_calib$plot
-spplot(AOA_calib$AOA$expected_RMSE,col.regions=viridis(100),main="expected RMSE")+
- spplot(AOA$AOA,col.regions=c("grey","transparent"))
+plot(AOA_calib$AOA$expected_RMSE,col=viridis(100),main="expected RMSE")
+plot(AOA$AOA,col=c("grey","transparent"),add=TRUE)
 
 ## ---- message = FALSE, warning=FALSE------------------------------------------
 dat <- get(load(system.file("extdata","Cookfarm.RData",package="CAST")))
@@ -169,9 +190,9 @@ dat <- aggregate(dat[,c("VW","Easting","Northing")],by=list(as.character(dat$SOU
 pts <- st_as_sf(dat,coords=c("Easting","Northing"))
 
 ##### Extract Predictors for the locations of the sampling points
-studyArea <- stack(system.file("extdata","predictors_2012-03-25.grd",package="CAST"))
+studyArea <- rast(system.file("extdata","predictors_2012-03-25.grd",package="CAST"))
 st_crs(pts) <- crs(studyArea)
-trainDat <- extract(studyArea,pts,df=TRUE)
+trainDat <- extract(studyArea,pts,na.rm=FALSE)
 pts$ID <- 1:nrow(pts)
 trainDat <- merge(trainDat,pts,by.x="ID",by.y="ID")
 # The final training dataset with potential predictors and VW:
@@ -188,17 +209,19 @@ model
 
 ## ---- message = FALSE, warning=FALSE------------------------------------------
 #Predictors:
-spplot(stretch(studyArea[[predictors]]))
+plot(stretch(studyArea[[predictors]]))
 
 #prediction:
-prediction <- predict(studyArea,model)
+prediction <- predict(studyArea,model,na.rm=TRUE)
 
-## ---- message = FALSE, warning=FALSE------------------------------------------
+## ---- message = FALSE, warning=FALSE,  fig.show="hold", out.width="50%"-------
 AOA <- aoa(studyArea,model)
 
 #### Plot results:
-grid.arrange(spplot(AOA$DI,col.regions=viridis(100),main="DI with sampling locations (red)")+
-               spplot(as_Spatial(pts),zcol="ID",col.regions="red"),
-  spplot(prediction, col.regions=viridis(100),main="prediction for AOA \n(LOOCV error applies)")+ spplot(AOA$AOA,col.regions=c("grey","transparent")),ncol=2)
+plot(AOA$DI,col=viridis(100),main="DI with sampling locations (red)")
+plot(pts,zcol="ID",col="red",add=TRUE)
+
+plot(prediction, col=viridis(100),main="prediction for AOA \n(LOOCV error applies)")
+plot(AOA$AOA,col=c("grey","transparent"),add=TRUE)